FIX :댓글 페이지네이션 API 적용g(#229)
상세페이지 백업 및 수정
댓글 수정 및 보완 페이지네이션 추가

closed #229

+++ src/components/units/users/myPage/mypage.queries.ts
@@ -38,7 +38,7 @@ export const POINT_HISTORY = gql`
   query fetchPointHistory(
     $startDate: DateTime
     $endDate: DateTime
-    $page: Float!
+    $page: Int!
   ) {
     fetchPointHistory(startDate: $startDate, endDate: $endDate, page: $page) {
       id
